+++ Engine/content/primitives.hpp
@@ -15,12 +15,12 @@
 
 namespace reveal3d::content {
 
-u16 GetCubeData(std::vector<render::Vertex> &vertices, std::vector<u16> &indices, u16 offset = 0);
-u16 GetPlaneData(std::vector<render::Vertex> &vertices, std::vector<u16> &indices, u16 offset = 0);
-u16 GetSphereData(std::vector<render::Vertex> &vertices, std::vector<u16> &indices, u16 offset = 0);
-u16 GetCylinderData(std::vector<render::Vertex> &vertices, std::vector<u16> &indices, u16 offset = 0);
-u16 GetConeData(std::vector<render::Vertex> &vertices, std::vector<u16> &indices, u16 offset = 0);
-u16 GetTorusData(std::vector<render::Vertex> &vertices, std::vector<u16> &indices, u16 offset = 0);
+u16 GetCubeData(std::vector<render::Vertex> &vertices, std::vector<u32> &indices, u16 offset = 0);
+u16 GetPlaneData(std::vector<render::Vertex> &vertices, std::vector<u32> &indices, u16 offset = 0);
+u16 GetSphereData(std::vector<render::Vertex> &vertices, std::vector<u32> &indices, u16 offset = 0);
+u16 GetCylinderData(std::vector<render::Vertex> &vertices, std::vector<u32> &indices, u16 offset = 0);
+u16 GetConeData(std::vector<render::Vertex> &vertices, std::vector<u32> &indices, u16 offset = 0);
+u16 GetTorusData(std::vector<render::Vertex> &vertices, std::vector<u32> &indices, u16 offset = 0);
 
 }
